added (float) for compile warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@5142 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtlib/vtopensg/NodeOSG.cpp
@@ -665,7 +665,7 @@ void vtTransform::Rotate2(const FPoint3 &axis, double angle)
 {
 	osg::Matrix4f mat = m_pTransform->getMatrix();
 
-	osg::Quaternion q(osg::Vec3f( axis.x, axis.y, axis.z), angle );
+	osg::Quaternion q(osg::Vec3f( axis.x, axis.y, axis.z), (float) angle );
 	osg::Matrix4f rot;  rot.setRotate(q);
 	mat.multLeft(rot);
 
@@ -678,7 +678,7 @@ void vtTransform::RotateLocal(const FPoint3 &axis, double angle)
 {
 	osg::Matrix4f mat = m_pTransform->getMatrix();
 
-	osg::Quaternion q(osg::Vec3f( axis.x, axis.y, axis.z), angle );
+	osg::Quaternion q(osg::Vec3f( axis.x, axis.y, axis.z), (float) angle );
 	osg::Matrix4f rot;  rot.setRotate(q);
 	mat.mult(rot);
 
@@ -703,7 +703,7 @@ void vtTransform::RotateParent(const FPoint3 &axis, double angle)
 	mat[3][2] = 0;
 
 	// rotate
-	osg::Quaternion q (osg::Vec3f( axis.x, axis.y, axis.z), angle);
+	osg::Quaternion q (osg::Vec3f( axis.x, axis.y, axis.z), (float) angle);
 	osg::Matrix4f rot; rot.setRotate(q);
 	mat.multLeft(rot);
